refactor(eval): split matcher.rs into module directory
